+++ python_modules/dagster-airflow/dagster_airflow/dagster_plugin.py
@@ -15,6 +15,7 @@ from airflow.utils.file import TemporaryDirectory
 from docker import APIClient, from_env, tls
 
 
+
 # This modified DockerOperator incorporates https://github.com/apache/airflow/pull/4315/files
 # for Docker compatibility on OS X, and additionally allows the Docker client to be configured
 # using `docker.from_env`.
